Standardize the way that memory values are handled (#104)
Lot of refactoring and code cleanup. Standardizing the way that memory values are dealt with, instead of converting back and forth from bytearrays/int/strings in multiple ways.

Makes a MemoryValue type which can be used to convert to and from different types. This requires a rewrite of a lot of things. Little to no changes are made to the logic and flow, however the interfaces and return types of many commonly used functions have been changed. A lot of the tests have to be rewritten for this new interface, but they are still testing the same functionality.

Adds tests for the MemoryValue type.