huge commit, many wip changes
rework the opcode_base class to be less dependant on assembly parameter. turns out
that Assembly Parameter will not be very useful for us as it wasn't implemented correctly
and has some design problems

instead, it would be better to continue with this pattern that is being used by OpCodeAdd
and related opcodes where each subclass will implement more shared functionality.
AssemblyParameter may not be necessary for now, may be more relevant when executing
but it doesn't handle imm right. it thinks that imm is in the op and not right after it.